requirements.txt
Django
djangorestframework
psycopg2-binary
celery
redis
django-celery-beat
pandas
pytest
pytest-django
django-filter
requests